+++ src/app/page.tsx
@@ -2,6 +2,8 @@ import MainContent from "@/contents/main";
 import { getTemplatesFromCDN } from "@/lib/helpers/data";
 import { toast } from "sonner";
 
+export const revalidate = 3600;
+
 export default async function Home() {
   const templates = await getTemplatesFromCDN()
   if(templates.error) toast.error(templates.error)