+++ src/lib/types/index.ts
@@ -1,4 +1,5 @@
 import { LucideIcon } from "lucide-react";
+import { FILTER_NAMES } from "../constants/filters";
 
 export interface SelectType{
      value: string,
@@ -6,12 +7,12 @@ export interface SelectType{
      Icon?: LucideIcon,
      disabled?: boolean
 }
-
+export type FilterName = keyof typeof FILTER_NAMES;
 export interface Template {
      slug: string;
      title: string;
      file: string;
      thumbnail: string;
-     categories: string[];
+     categories: FilterName[];
      pages: number;
 }