+++ src/lib/types/schemas.ts
@@ -1,0 +1,4 @@
+import * as z from "zod"
+import { SearchSchema } from "../schemas"
+
+export type SearchType = z.infer<typeof SearchSchema>